requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
configparser==5.3.0
greenlet==1.1.3.post0
idna==3.4
iniconfig==1.1.1
joblib==1.2.0
mysql-connector-python==8.0.31
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandasql==0.7.3
pluggy==1.0.0
presto-python-client==0.8.3
progressbar2==4.1.1
protobuf==3.20.1
psycopg2-binary==2.9.4
py==1.11.0
py4j==0.10.9.5
pyparsing==3.0.9
pyspark==3.3.1
pytest==7.1.3
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.5
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.3
six==1.16.0
SQLAlchemy==1.4.42
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.1
typing_extensions==4.4.0
urllib3==1.26.12
